create separate constituencies-map component
