+++ src/app/india-elections-2024/components/constituencies-map/constituencies-map.component.html
@@ -1,0 +1,1 @@
+<div class="map-container"></div>
