+++ src/app/india-elections-2024/components/overall-result/overall-result.component.html
@@ -1,4 +1,4 @@
 <h3>Overall Results</h3>
 <div class="overall-result-viz-container">
-  <div class="map-container"></div>
+  <app-constituencies-map />
 </div>
