+++ src/app/india-elections-2024/india-elections-2024.component.ts
@@ -1,5 +1,5 @@
-import { Component } from '@angular/core';
-import { OverallResultComponent } from "./overall-result/overall-result.component";
+import { Component, inject, OnInit } from '@angular/core';
+import { OverallResultComponent } from "./components/overall-result/overall-result.component";
 import { ElectionDataService } from "./services/election-data.service";
 import { ElectionDataStore } from "./election-data.store";
 
@@ -13,6 +13,11 @@ import { ElectionDataStore } from "./election-data.store";
   templateUrl: './india-elections-2024.component.html',
   styleUrl: './india-elections-2024.component.scss'
 })
-export class IndiaElections2024Component {
+export class IndiaElections2024Component implements OnInit {
+  private electionDataStore = inject(ElectionDataStore);
+
+  ngOnInit(): void {
+    this.electionDataStore.loadAllData();
+  }
 
 }
